# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "version": "0.1"
    },
    "paths": {
        "/api/agentes": {
            "get": {
                "tags": [
                    "Agente"
                ],
                "summary": "getAgenteList",
                "description": "Get all Agentes",
                "operationId": "8deeff578f5d2d74b975be976f80d44c",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agente"
                ],
                "summary": "createAgente",
                "description": "Create Agente",
                "operationId": "d4fe17793707cc26016137ad51e15b7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agentes/{id}": {
            "get": {
                "tags": [
                    "Agente"
                ],
                "summary": "getAgenteItem",
                "description": "Get Agente",
                "operationId": "0ad206ba83a91d97aec4ab35e0d8728f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Agente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agente"
                ],
                "summary": "updateAgente",
                "description": "Update Agente",
                "operationId": "a4be971022406fb8329a66147a29f7e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Agente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agente"
                ],
                "summary": "deleteAgente",
                "description": "Delete Agente",
                "operationId": "8f29f6a36175d234dc3345e6601431ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Agente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    }
}